+++ examples/v3_0/src/components/schemas/ref_single_all_of.yaml
